Fix level Frog Pond (for-pond.stl) to be beatable
To complete this level you are required to stand on forest rocks. The intended behavior is that they are solid, however, properties of the rocks have changed at some point in time, so the rocks are not solid, making the level unbeatable. I've added a few invisible blocks to an additional layer to make the rocks behave as intended.